requirements.txt
annotated-types==0.6.0
apache-airflow-providers-common-io==1.2.0
apache-airflow-providers-common-sql==1.10.1
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-http==4.9.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-sqlite==3.7.0
apispec==6.4.0
appdirs==1.4.4
argcomplete==3.2.2
asgiref==3.7.2
atomicwrites==1.4.0
attrs==23.2.0
backoff==2.2.1
bar==0.2.1
blinker==1.7.0
Brotli==1.1.0
brotlipy==0.7.0
cachelib==0.9.0
cachetools==5.3.2
chardet==3.0.4
clickclick==20.10.2
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
constantly==15.1.0
cron-descriptor==1.4.3
croniter==2.0.1
cssselect==1.1.0
daal4py==2023.1.1
decorator==4.4.2
Deprecated==1.2.14
distro==1.8.0
et-xmlfile==1.1.0
fonttools==4.25.0
fpdf==1.7.2
fsspec==2024.2.0
# google-ai-generativelanguage==0.4.0S
# google-api-core==2.14.0
# google-api-python-client==2.109.0
# google-auth==2.23.4
# google-auth-httplib2==0.1.1
# google-auth-oauthlib==1.1.0
# google-generativeai==0.3.2
# google-re2==1.1
# googleapis-common-protos==1.61.0
# googletrans==3.0.0
grpcio==1.60.0
grpcio-status==1.60.0
gspread==5.12.4
gspread-dataframe==3.3.1 
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.22.0
# httpx#==0.13.3 #0.23.0
hyperframe==5.2.0
idna==2.10
imageio-ffmpeg==0.4.9
importlib-resources==6.1.1
inflate64==1.0.0
inflection==0.5.1
instaloader==4.10.2
jsonpointer==2.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
limits==3.7.0
lockfile==0.12.2
loguru==0.7.2
marshmallow==3.20.2
marshmallow-oneofschema==3.1.0
marshmallow-sqlalchemy==0.26.1
moviepy==1.0.3
multivolumefile==0.2.3
munkres==1.1.4
oauthlib==3.2.2
# openai==1.14.2
openpyxl==3.1.2
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
ordered-set==4.1.0
pandas==2.2.0
patsy==0.5.3
pendulum==3.0.0
pep8==1.7.1
ply==3.11
prison==0.2.1
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.21.12
psycopg2-binary==2.9.9
py7zr==0.20.8
pyarrow==11.0.0
pyasn1-modules==0.2.8
pybcj==1.0.2
pycryptodomex==3.20.0
pydantic==2.6.0
pydantic_core==2.16.1
PyDispatcher==2.0.5
pydub==0.25.1
pyls-spyder==0.4.0
PyMySQL==1.1.0
pyppmd==1.1.0
python-lsp-jsonrpc==1.0.0
pyzstd==0.15.9
queuelib==1.5.0
ratelimit==2.2.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scipy==1.11.1
scrapfly-sdk==0.8.10
selenium==4.16.0
setproctitle==1.3.3
sidrapy==0.1.4
sniffio==1.3.0
SpeechRecognition==3.10.1
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
termcolor==2.4.0
textblob==0.17.1
texttable==1.7.0
time-machine==2.13.0
trio==0.23.1
trio-websocket==0.11.1
unicodecsv==0.14.1
universal-pathlib==0.1.4
uritemplate==4.1.1
urllib3==2.2.1
webdriver-manager==4.0.1
webencodings==0.5.1
wsproto==1.2.0
WTForms==3.1.2
XlsxWriter==3.1.9
zstandard==0.19.0